+++ src/util/__init__.py
@@ -3,4 +3,4 @@ from .validator import is_valid_phone, is_valid_student_id, is_valid_year, is_va
 from .get_scsc_global_status import get_scsc_global_status
 from .create_uuid import create_uuid
 from .get_file_extension import get_file_extension
-from .get_user_role_level import get_user_role_level
+from .user_role import get_user_role_level
